Pentagram v0.0.1: initial review of foundation
Pentagram is an initial import and partial rewrite of Clio's internal daemon.rb
library. Fundamentally it is largely the same, but bits and pieces of it have
been modified in order to make it less "Clio-centric" (i.e., Log4r has been
removed, ProcessHandle has been removed, and there have been various tweaks
along the way to make it more generally useful in a public sense).

Initial changes have been made as suggested by Dennis Taylor, but it is expected
that further changes will be required in order to get this library "up to
snuff".

As it stands Pentagram v0.0.1 is completely untested, and not ready for any
sort of use.