Fix signal handling regression caused by refactor
In the changes suggested during the initial review, we changed @@signal_handlers
to be `Hash.new { Array.new }` by default, which meant that our code for
registering new signal handlers in `Daemon.register_signal_handler` stopped
working due to the handler being added to an array object that was created and
then thrown away.

This commit reverts this change in order to fix the regression.

Additionally, protect against accidental registration of a signal handler that
does not have the expected arity by checking signal handler arity during
registration.