Pentagram v1.0.0: Call it Stable
The rewrite of Pentagram has undergone two reviews, in-house testing, and open
source review as well. It looks reasonably good, so we are going to cut v1.0.0
instead of getting permanently stuck with version 0.X.Y because it is "never
ready".

We claim that this version is ready for legitimate attempts to use it.