Pentagram v2.1.1: do not handle SIGCHLD by default.
Explicit handling of SIGCHLD is, I believe, a comparatively rare need in the
present day-and-age. Generally speaking, I expect that most uses of Pentagram
will _not_ want their daemon to be interrupted in order to deal with SIGCHLD.

Ergo, we stop catching it by default and add a comment describing to the reader
how they should re-enable framework handling of it if they desire to do so.