Pentagram v2.1.2: do not handle SIGWINCH by default.
Similar to SIGCHLD, explicit handling of SIGWINCH is a comparatively rare need,
particularly in the field of daemons or background processes. Generally
speaking, most daemons or background processes do not want to handle SIGWINCH
nor be informed when the 'window' changes size.

Ergo, we apply the same behaviour change as we did for SIGCHLD: we stop catching
it by default.